Task-43673: Refactoring code news details and news apps
