+++ webapp/src/main/webapp/css/news.less
@@ -20,19 +20,19 @@
     margin: 1.5rem;
   }
   .shareNewsDrawer {
-    .share-news {
-      .newsInformationCard {
+    .share-activity {
+      .activityInformationCard {
         max-width: ~"calc(100% - 20px)";
       }
 
-      .newsShareDescription {
+      .activityShareDescription {
         min-width: ~"calc(100% - 28px)";
         border-radius: 5px;
         box-shadow: none;
         resize: none;
       }
 
-      .newsSpaceSuggester {
+      .activitySpaceSuggester {
         max-width: ~"calc(100% - 20px)";
         border-radius: 5px;
       }
