+++ webapp/src/main/webapp/news-details/initComponents.js
@@ -2,18 +2,12 @@ import ExoNewsArchive  from './components/ExoNewsArchive.vue';
 import ExoNewsDetails  from './components/ExoNewsDetails.vue';
 import ExoNewsDetailsActionMenu from './components/ExoNewsDetailsActionMenu.vue';
 import ExoNewsPin from './components/ExoNewsPin.vue';
-import ExoNewsShareActivityDrawer from './components/ExoNewsShareActivityDrawer.vue';
-import ExoNewsSpaceItem from './components/ExoNewsSpaceItem.vue';
-import ExoNewsSpaceSuggester from './components/ExoNewsSpaceSuggester.vue';
 
 const components = {
   'exo-news-archive': ExoNewsArchive,
   'exo-news-details': ExoNewsDetails,
   'exo-news-details-action-menu': ExoNewsDetailsActionMenu,
   'exo-news-pin': ExoNewsPin,
-  'exo-news-share-activity-drawer': ExoNewsShareActivityDrawer,
-  'exo-news-space-item': ExoNewsSpaceItem,
-  'exo-news-space-suggester': ExoNewsSpaceSuggester,
 };
 
 for (const key in components) {
