+++ webapp/src/main/webapp/news/components/NewsApp.vue
@@ -157,7 +157,13 @@
     <div v-if="showLoadMoreButton" class="newsListPagination">
       <div class="btn btn-block" @click="loadMore">{{ $t('news.app.loadMore') }}</div>
     </div>
-    <exo-news-share-activity-drawer />
+    <v-app>
+      <share-activity-drawer
+        ref="shareNewsDrawer"
+        class="shareNewsDrawer"
+        :activity-id="activityId"
+        @share-activity="shareNews" />
+    </v-app>
     <news-activity-sharing-spaces-drawer />
     <exo-news-notification-alerts />
   </div>
@@ -184,6 +190,7 @@ export default {
       showArchiveButton: true,
       loadingNews: true,
       initialized: false,
+      activityId: null
     };
   },
   computed: {
@@ -253,6 +260,11 @@ export default {
       if (news && news.spaceId) {
         this.fetchNews(false);
       }
+    });
+    this.$root.$on('news-share-drawer-open', news => {
+      this.news = news;
+      this.activityId = news.activityId;
+      this.open();
     });
   },
   methods: {
@@ -395,7 +407,28 @@ export default {
           this.fetchNews(false);
         }
       }, redirectionTime);
-    }
-  }
+    },
+    open() {
+      if (this.$refs.shareNewsDrawer) {
+        this.$refs.shareNewsDrawer.open();
+      }
+    },
+    close() {
+      this.$refs.shareNewsDrawer.close();
+    },
+    shareNews(spaces, description) {
+      const spacesList = [];
+      spaces.forEach(space => {
+        this.$spaceService.getSpaceByPrettyName(space,'identity').then(data => {
+          spacesList.push(data.displayName);
+        });
+      });
+      this.$newsServices.shareNews(this.news.newsId, this.news.activityId, description, spaces)
+        .then(() => {
+          this.$root.$emit('news-shared', this.news, spacesList);
+          this.close();
+        });
+    },
+  },
 };
 </script>
